Texture loading improvement
- use texture image data for hash generation
- use std::move for pending texture clearing in TextureDataManager
- do not add the texture to be loaded again when the status changes

+++ src/render/texture/texture_p.h
@@ -73,7 +73,7 @@ class TextureManager;
 class TextureImageManager;
 class TextureDataManager;
 
-typedef uint TextureDNA;
+typedef quint64 TextureDNA;
 
 class Texture : public BackendNode
 {
